requirements.txt
eventlet
python-socketio
requests
flask